@@ -5,37 +5,45 @@
   "author": {
     "name": "fiahfy",
     "email": "d.fiahfy@gmail.com",
-    "url": "http://fiahfy.blogspot.com/"
+    "url": "https://fiahfy.github.io/"
   },
   "dependencies": {
-    "@fiahfy/electron-accelerator-formatter": "^0.0.2",
-    "@fiahfy/electron-context-menu": "^0.0.2",
-    "@fiahfy/simple-clone": "^0.0.2",
-    "@nuxtjs/vuetify": "^0.5.1",
-    "d3": "^5.7.0",
+    "@fiahfy/electron-accelerator-formatter": "^0.0.4",
+    "@fiahfy/electron-context-menu": "^0.0.6",
+    "@fiahfy/simple-clone": "^0.0.3",
+    "@mdi/font": "^5.3.45",
+    "@nuxtjs/vuetify": "^1.11.2",
+    "@vue/composition-api": "^0.6.7",
+    "d3": "^5.16.0",
     "debounce": "^1.2.0",
-    "electron-devtools-installer": "^2.2.0",
-    "electron-window-state": "^5.0.1",
-    "material-design-icons-iconfont": "^5.0.1",
-    "nuxt": "^2.3.4",
-    "typeface-roboto": "^0.0.54",
-    "vuex-persistedstate": "^2.0.0"
+    "electron-devtools-installer": "^3.0.0",
+    "electron-window-state": "^5.0.3",
+    "nuxt": "^2.13.1",
+    "pretty-bytes": "^5.3.0",
+    "typeface-roboto": "^0.0.75",
+    "vuex-module-decorators": "^0.17.0",
+    "vuex-persistedstate": "^3.0.1"
   },
   "devDependencies": {
-    "@fiahfy/icns-convert": "^0.0.5",
-    "@fiahfy/ico-convert": "^0.0.4",
-    "cross-env": "^5.0.5",
-    "electron": "^5.0.2",
-    "electron-builder": "^20.0.4",
-    "eslint": "^5.11.0",
-    "eslint-config-prettier": "^4.0.0",
-    "eslint-plugin-prettier": "^3.0.0",
-    "eslint-plugin-vue": "^5.0.0-beta.3",
-    "node-sass": "^4.9.0",
-    "npm-run-all": "^4.0.2",
-    "prettier": "^1.14.3",
-    "sass-loader": "^7.0.1",
-    "vuetify-loader": "^1.1.1",
+    "@fiahfy/icns-convert": "^0.0.8",
+    "@fiahfy/ico-convert": "^0.0.7",
+    "@nuxt/typescript-build": "^1.0.3",
+    "@nuxtjs/eslint-config-typescript": "^2.0.0",
+    "@types/d3": "^5.7.2",
+    "@types/debounce": "^1.2.0",
+    "@types/node": "^12.12.47",
+    "cross-env": "^7.0.2",
+    "electron": "^8.3.3",
+    "electron-builder": "^22.7.0",
+    "eslint": "^7.3.1",
+    "eslint-config-prettier": "^6.11.0",
+    "eslint-plugin-prettier": "^3.1.4",
+    "husky": "^4.2.5",
+    "node-sass": "^4.14.1",
+    "npm-run-all": "^4.1.5",
+    "prettier": "^2.0.5",
+    "sass-loader": "^8.0.2",
+    "vuetify-loader": "^1.5.0",
     "worker-loader": "^2.0.0"
   },
   "keywords": [
@@ -55,8 +63,8 @@
     "build": "nuxt generate",
     "dev": "cross-env NODE_ENV=development run-p serve electron",
     "electron": "electron main.js",
-    "lint": "eslint -f codeframe --ext .js,.vue --ignore-path .gitignore .",
-    "release": "build -mw",
+    "lint": "eslint -f codeframe --ext .ts,.vue --ignore-path .gitignore .",
+    "release": "electron-builder",
     "resources": "run-p resources:*",
     "resources:icns": "icns-convert build/icon.png",
     "resources:ico": "ico-convert build/icon.png",
